create; favorites list logic
