+++ project/src/index.js
@@ -8,7 +8,6 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
-/* import { createStore, applyMiddleware } from 'redux'; */
 import { configureStore } from '@reduxjs/toolkit';
 import thunk from 'redux-thunk';
 import App from './components/app/app';
